deaddrop-api/requirements.txt
cffi==1.10.0
cryptography==1.7.2
Django==1.11.3
django-cors-headers==2.1.0
django-extensions==1.8.1
djangorestframework==3.6.3
idna==2.5
pyasn1==0.2.3
pycparser==2.18
PyJWT==1.5.2
pytz==2017.2
six==1.10.0
